requirements.txt
backcall==0.1.0
chardet==3.0.4
decorator==4.3.0
Django==2.1.1
django-ckeditor==5.6.1
django-js-asset==1.1.0
ipython==6.5.0
ipython-genutils==0.2.0
jedi==0.12.1
parso==0.3.1
pexpect==4.6.0
pickleshare==0.7.4
Pillow==5.2.0
prompt-toolkit==1.0.15
ptyprocess==0.6.0
Pygments==2.2.0
pypugjs==5.7.2
pytz==2018.5
simplegeneric==0.8.1
six==1.11.0
traitlets==4.3.2
wcwidth==0.1.7
